Adjust FileIO to compile
