+++ Library/Utilities/FileIO.hpp
@@ -32,27 +32,12 @@
 
 // Standard C++ header files
 #include <string>
-#include <sstream>
 #include <vector>
 
 namespace solutio
 {
-  void LineRead(const std::string &s, char delim, std::vector<std::string> &elems)
-  {
-    std::stringstream ss;
-    ss.str(s);
-    std::string item;
-    while (std::getline(ss, item, delim)){
-      elems.push_back(item);
-    }
-  }
-  
-  std::vector<std::string> LineRead(const std::string &s, char delim)
-  {
-    std::vector<std::string> elems;
-    LineRead(s, delim, elems);
-    return elems;
-  }
+  void LineRead(const std::string &s, char delim, std::vector<std::string> &elems);
+  std::vector<std::string> LineRead(const std::string &s, char delim);
 }
 
 // End header guard
